Extract domain into ferriskey-domain crate
Move client/role/user/common/authentication domain types, value
objects, ports and commands into a new libs/ferriskey-domain crate.
refactor: introduce associated types (Error, Identity) on service and
policy
traits and set concrete implementations in core (e.g. `type Error =
CoreError`, `type Identity = Identity`). Replace duplicated domain
definitions in core with re-exports from ferriskey-domain, add test
mocks, and update repositories/services to use the new error type.
Add thiserror dependency for domain crate.
