+++ core/src/domain/role/ports.rs
@@ -1,108 +1,59 @@
-use uuid::Uuid;
+pub use ferriskey_domain::role::ports::{RolePolicy, RoleRepository, RoleService};
 
-use crate::domain::realm::entities::RealmId;
-use crate::domain::{
-    authentication::value_objects::Identity,
-    common::entities::app_errors::CoreError,
-    realm::entities::Realm,
-    role::{
-        entities::{GetUserRolesInput, Role, UpdateRoleInput},
+#[cfg(test)]
+pub use mocks::MockRoleRepository;
+
+#[cfg(test)]
+mod mocks {
+    use mockall::mock;
+    use uuid::Uuid;
+
+    use crate::domain::common::entities::app_errors::CoreError;
+    use ferriskey_domain::realm::RealmId;
+    use ferriskey_domain::role::{
+        entities::Role,
         value_objects::{CreateRoleRequest, UpdateRolePermissionsRequest, UpdateRoleRequest},
-    },
-};
+    };
 
-pub trait RoleService: Send + Sync {
-    fn delete_role(
-        &self,
-        identity: Identity,
-        realm_name: String,
-        role_id: Uuid,
-    ) -> impl Future<Output = Result<(), CoreError>> + Send;
-    fn get_role(
-        &self,
-        identity: Identity,
-        realm_name: String,
-        role_id: Uuid,
-    ) -> impl Future<Output = Result<Role, CoreError>> + Send;
-    fn get_roles(
-        &self,
-        identity: Identity,
-        realm_name: String,
-    ) -> impl Future<Output = Result<Vec<Role>, CoreError>> + Send;
-    fn update_role_permissions(
-        &self,
-        identity: Identity,
-        realm_name: String,
-        role_id: Uuid,
-        permissions: Vec<String>,
-    ) -> impl Future<Output = Result<Role, CoreError>> + Send;
-    fn update_role(
-        &self,
-        identity: Identity,
-        input: UpdateRoleInput,
-    ) -> impl Future<Output = Result<Role, CoreError>> + Send;
-    fn get_user_roles(
-        &self,
-        identity: Identity,
-        input: GetUserRolesInput,
-    ) -> impl Future<Output = Result<Vec<Role>, CoreError>> + Send;
-}
+    mock! {
+        pub RoleRepository {}
+        impl ferriskey_domain::role::ports::RoleRepository for RoleRepository {
 
-pub trait RolePolicy: Send + Sync {
-    fn can_create_role(
-        &self,
-        identity: &Identity,
-        target_realm: &Realm,
-    ) -> impl Future<Output = Result<bool, CoreError>> + Send;
-    fn can_view_role(
-        &self,
-        identity: &Identity,
-        target_realm: &Realm,
-    ) -> impl Future<Output = Result<bool, CoreError>> + Send;
-    fn can_update_role(
-        &self,
-        identity: &Identity,
-        target_realm: &Realm,
-    ) -> impl Future<Output = Result<bool, CoreError>> + Send;
-    fn can_delete_role(
-        &self,
-        identity: &Identity,
-        target_realm: &Realm,
-    ) -> impl Future<Output = Result<bool, CoreError>> + Send;
-}
+            fn create(
+                &self,
+                payload: CreateRoleRequest,
+            ) -> impl Future<Output = Result<Role, CoreError>> + Send;
+            fn get_by_client_id(
+                &self,
+                client_id: Uuid,
+            ) -> impl Future<Output = Result<Vec<Role>, CoreError>> + Send;
+            fn get_by_id(
+                &self,
+                id: Uuid,
+            ) -> impl Future<Output = Result<Option<Role>, CoreError>> + Send;
+            fn delete_by_id(&self, id: Uuid) -> impl Future<Output = Result<(), CoreError>> + Send;
 
-#[cfg_attr(test, mockall::automock)]
-pub trait RoleRepository: Send + Sync {
-    fn create(
-        &self,
-        payload: CreateRoleRequest,
-    ) -> impl Future<Output = Result<Role, CoreError>> + Send;
-    fn get_by_client_id(
-        &self,
-        client_id: Uuid,
-    ) -> impl Future<Output = Result<Vec<Role>, CoreError>> + Send;
-    fn get_by_id(&self, id: Uuid) -> impl Future<Output = Result<Option<Role>, CoreError>> + Send;
-    fn delete_by_id(&self, id: Uuid) -> impl Future<Output = Result<(), CoreError>> + Send;
+            fn find_by_realm_id(
+                &self,
+                realm_id: RealmId,
+            ) -> impl Future<Output = Result<Vec<Role>, CoreError>> + Send;
+            fn find_by_name(
+                &self,
+                name: String,
+                realm_id: Uuid,
+            ) -> impl Future<Output = Result<Option<Role>, CoreError>> + Send;
 
-    fn find_by_realm_id(
-        &self,
-        realm_id: RealmId,
-    ) -> impl Future<Output = Result<Vec<Role>, CoreError>> + Send;
-    fn find_by_name(
-        &self,
-        name: String,
-        realm_id: Uuid,
-    ) -> impl Future<Output = Result<Option<Role>, CoreError>> + Send;
+            fn update_by_id(
+                &self,
+                id: Uuid,
+                payload: UpdateRoleRequest,
+            ) -> impl Future<Output = Result<Role, CoreError>> + Send;
 
-    fn update_by_id(
-        &self,
-        id: Uuid,
-        payload: UpdateRoleRequest,
-    ) -> impl Future<Output = Result<Role, CoreError>> + Send;
-
-    fn update_permissions_by_id(
-        &self,
-        id: Uuid,
-        payload: UpdateRolePermissionsRequest,
-    ) -> impl Future<Output = Result<Role, CoreError>> + Send;
+            fn update_permissions_by_id(
+                &self,
+                id: Uuid,
+                payload: UpdateRolePermissionsRequest,
+            ) -> impl Future<Output = Result<Role, CoreError>> + Send;
+        }
+    }
 }
