+++ core/src/domain/user/entities.rs
@@ -1,167 +1,2 @@
-use std::fmt::Display;
-
-use chrono::{DateTime, Utc};
-use serde::{Deserialize, Serialize};
-use thiserror::Error;
-use utoipa::ToSchema;
-use uuid::Uuid;
-
-use crate::domain::realm::entities::RealmId;
-use crate::domain::{common::generate_uuid_v7, realm::entities::Realm, role::entities::Role};
-
-#[derive(Debug, Clone, Deserialize, Serialize, ToSchema, PartialEq)]
-pub struct User {
-    pub id: Uuid,
-    pub realm_id: RealmId,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub client_id: Option<Uuid>,
-    pub username: String,
-    pub firstname: String,
-    pub lastname: String,
-    pub email: String,
-    pub email_verified: bool,
-    pub enabled: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub roles: Option<Vec<Role>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub realm: Option<Realm>,
-    pub required_actions: Vec<RequiredAction>,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-}
-
-pub struct UserConfig {
-    pub realm_id: RealmId,
-    pub client_id: Option<Uuid>,
-    pub username: String,
-    pub firstname: String,
-    pub lastname: String,
-    pub email: String,
-    pub email_verified: bool,
-    pub enabled: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord, Serialize, Deserialize, ToSchema)]
-pub enum RequiredAction {
-    #[serde(rename = "configure_otp")]
-    ConfigureOtp,
-
-    #[serde(rename = "verify_email")]
-    VerifyEmail,
-
-    #[serde(rename = "update_password")]
-    UpdatePassword,
-}
-
-impl Display for RequiredAction {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            RequiredAction::ConfigureOtp => write!(f, "configure_otp"),
-            RequiredAction::VerifyEmail => write!(f, "verify_email"),
-            RequiredAction::UpdatePassword => write!(f, "update_password"),
-        }
-    }
-}
-
-impl TryFrom<String> for RequiredAction {
-    type Error = RequiredActionError;
-
-    fn try_from(value: String) -> Result<Self, Self::Error> {
-        match value.as_str() {
-            "configure_otp" => Ok(RequiredAction::ConfigureOtp),
-            "verify_email" => Ok(RequiredAction::VerifyEmail),
-            "update_password" => Ok(RequiredAction::UpdatePassword),
-            _ => Err(RequiredActionError::Invalid),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Error)]
-pub enum RequiredActionError {
-    #[error("Required action not found")]
-    NotFound,
-    #[error("Required action already exists")]
-    AlreadyExists,
-    #[error("Invalid required action")]
-    Invalid,
-    #[error("Internal server error")]
-    InternalServerError,
-}
-
-impl User {
-    pub fn new(user_config: UserConfig) -> Self {
-        let now = Utc::now();
-        let id = generate_uuid_v7();
-
-        Self {
-            id,
-            realm_id: user_config.realm_id,
-            client_id: user_config.client_id,
-            username: user_config.username,
-            firstname: user_config.firstname,
-            lastname: user_config.lastname,
-            email: user_config.email,
-            email_verified: user_config.email_verified,
-            enabled: user_config.enabled,
-            roles: Some(Vec::new()),
-            realm: None,
-            required_actions: Vec::new(),
-            created_at: now,
-            updated_at: now,
-        }
-    }
-}
-
-pub struct ResetPasswordInput {
-    pub user_id: Uuid,
-    pub password: String,
-    pub temporary: bool,
-    pub realm_name: String,
-}
-
-pub struct AssignRoleInput {
-    pub realm_name: String,
-    pub user_id: Uuid,
-    pub role_id: Uuid,
-}
-
-pub struct BulkDeleteUsersInput {
-    pub realm_name: String,
-    pub ids: Vec<Uuid>,
-}
-
-pub struct CreateUserInput {
-    pub realm_name: String,
-    pub username: String,
-    pub firstname: String,
-    pub lastname: String,
-    pub email: String,
-    pub email_verified: Option<bool>,
-}
-
-pub struct GetUserInput {
-    pub realm_name: String,
-    pub user_id: Uuid,
-}
-
-pub struct UpdateUserInput {
-    pub realm_name: String,
-    pub user_id: Uuid,
-    pub firstname: String,
-    pub lastname: String,
-    pub email: String,
-    pub email_verified: Option<bool>,
-    pub enabled: bool,
-    pub required_actions: Option<Vec<String>>,
-}
-
-pub struct UnassignRoleInput {
-    pub realm_name: String,
-    pub user_id: Uuid,
-    pub role_id: Uuid,
-}
-
-pub struct GetUserPermissionsInput {
-    pub realm_name: String,
-    pub user_id: Uuid,
-}
+pub use ferriskey_domain::user::commands::*;
+pub use ferriskey_domain::user::entities::*;
