+++ libs/ferriskey-domain/src/authentication/mod.rs
@@ -1,0 +1,1 @@
+pub mod entities;
