+++ libs/ferriskey-domain/src/client/entities/redirect_uri.rs
@@ -1,7 +1,9 @@
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
-use uuid::{NoContext, Timestamp, Uuid};
+use uuid::Uuid;
+
+use crate::generate_timestamp;
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Deserialize, Serialize, ToSchema)]
 pub struct RedirectUri {
@@ -15,10 +17,7 @@ pub struct RedirectUri {
 
 impl RedirectUri {
     pub fn new(client_id: Uuid, value: String, enabled: bool) -> Self {
-        let now = Utc::now();
-        let seconds = now.timestamp().try_into().unwrap_or(0);
-
-        let timestamp = Timestamp::from_unix(NoContext, seconds, 0);
+        let (now, timestamp) = generate_timestamp();
 
         Self {
             id: Uuid::new_v7(timestamp),
