+++ libs/ferriskey-domain/src/role/mod.rs
@@ -1,0 +1,4 @@
+pub mod commands;
+pub mod entities;
+pub mod ports;
+pub mod value_objects;
