+++ libs/ferriskey-domain/src/role/value_objects.rs
@@ -1,0 +1,24 @@
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+
+use crate::realm::RealmId;
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct CreateRoleRequest {
+    pub name: String,
+    pub description: Option<String>,
+    pub permissions: Vec<String>,
+    pub realm_id: RealmId,
+    pub client_id: Option<Uuid>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct UpdateRoleRequest {
+    pub name: Option<String>,
+    pub description: Option<String>,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct UpdateRolePermissionsRequest {
+    pub permissions: Vec<String>,
+}
