+++ libs/ferriskey-domain/src/user/entities.rs
@@ -1,0 +1,114 @@
+use std::fmt::Display;
+
+use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
+use thiserror::Error;
+use utoipa::ToSchema;
+use uuid::Uuid;
+
+use crate::generate_uuid_v7;
+use crate::realm::{Realm, RealmId};
+use crate::role::entities::Role;
+
+#[derive(Debug, Clone, Deserialize, Serialize, ToSchema, PartialEq)]
+pub struct User {
+    pub id: Uuid,
+    pub realm_id: RealmId,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub client_id: Option<Uuid>,
+    pub username: String,
+    pub firstname: String,
+    pub lastname: String,
+    pub email: String,
+    pub email_verified: bool,
+    pub enabled: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub roles: Option<Vec<Role>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub realm: Option<Realm>,
+    pub required_actions: Vec<RequiredAction>,
+    pub created_at: DateTime<Utc>,
+    pub updated_at: DateTime<Utc>,
+}
+
+pub struct UserConfig {
+    pub realm_id: RealmId,
+    pub client_id: Option<Uuid>,
+    pub username: String,
+    pub firstname: String,
+    pub lastname: String,
+    pub email: String,
+    pub email_verified: bool,
+    pub enabled: bool,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord, Serialize, Deserialize, ToSchema)]
+pub enum RequiredAction {
+    #[serde(rename = "configure_otp")]
+    ConfigureOtp,
+
+    #[serde(rename = "verify_email")]
+    VerifyEmail,
+
+    #[serde(rename = "update_password")]
+    UpdatePassword,
+}
+
+impl Display for RequiredAction {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            RequiredAction::ConfigureOtp => write!(f, "configure_otp"),
+            RequiredAction::VerifyEmail => write!(f, "verify_email"),
+            RequiredAction::UpdatePassword => write!(f, "update_password"),
+        }
+    }
+}
+
+impl TryFrom<String> for RequiredAction {
+    type Error = RequiredActionError;
+
+    fn try_from(value: String) -> Result<Self, Self::Error> {
+        match value.as_str() {
+            "configure_otp" => Ok(RequiredAction::ConfigureOtp),
+            "verify_email" => Ok(RequiredAction::VerifyEmail),
+            "update_password" => Ok(RequiredAction::UpdatePassword),
+            _ => Err(RequiredActionError::Invalid),
+        }
+    }
+}
+
+#[derive(Debug, Clone, Error)]
+pub enum RequiredActionError {
+    #[error("Required action not found")]
+    NotFound,
+    #[error("Required action already exists")]
+    AlreadyExists,
+    #[error("Invalid required action")]
+    Invalid,
+    #[error("Internal server error")]
+    InternalServerError,
+}
+
+impl User {
+    pub fn new(user_config: UserConfig) -> Self {
+        let now = Utc::now();
+        let id = generate_uuid_v7();
+
+        Self {
+            id,
+            realm_id: user_config.realm_id,
+            client_id: user_config.client_id,
+            username: user_config.username,
+            firstname: user_config.firstname,
+            lastname: user_config.lastname,
+            email: user_config.email,
+            email_verified: user_config.email_verified,
+            enabled: user_config.enabled,
+            roles: Some(Vec::new()),
+            realm: None,
+            required_actions: Vec::new(),
+            created_at: now,
+            updated_at: now,
+        }
+    }
+}
