+++ libs/ferriskey-domain/src/user/mod.rs
@@ -1,0 +1,3 @@
+pub mod commands;
+pub mod entities;
+pub mod value_objects;
